Refactor BGRA to RGBA function.
