+++ src/isyntax_example2.c
@@ -1,5 +1,4 @@
 #include "libisyntax.h"
-
 #include <stdint.h>
 #include <assert.h>
 
@@ -7,54 +6,8 @@
 
 #include "third_party/stb_image_write.h"  // for png export
 
-#if defined(__ARM_NEON)
-
-#include <arm_neon.h>
-
-#elif defined(__SSE2__)
-#include <emmintrin.h>
-#endif
-
-
 #define LOG_VAR(fmt, var) printf("%s: %s=" fmt "\n", __FUNCTION__, #var, var)
 
-void bgra_to_rgba(uint32_t *pixels, int tile_width, int tile_height) {
-    int num_pixels = tile_width * tile_height;
-    int num_pixels_aligned = (num_pixels / 4) * 4;
-
-#if defined(__ARM_NEON)
-    for (int i = 0; i < num_pixels_aligned; i += 4) {
-        uint32x4_t bgra = vld1q_u32(pixels + i);
-        uint32x4_t b_mask = vdupq_n_u32(0x000000FF);
-        uint32x4_t r_mask = vdupq_n_u32(0x00FF0000);
-        uint32x4_t b = vandq_u32(bgra, b_mask);
-        uint32x4_t r = vandq_u32(bgra, r_mask);
-        uint32x4_t br_swapped = vorrq_u32(vshlq_n_u32(b, 16), vshrq_n_u32(r, 16));
-        uint32x4_t ga_alpha_mask = vdupq_n_u32(0xFF00FF00);
-        uint32x4_t ga_alpha = vandq_u32(bgra, ga_alpha_mask);
-        uint32x4_t rgba = vorrq_u32(ga_alpha, br_swapped);
-        vst1q_u32(pixels + i, rgba);
-    }
-#elif defined(__SSE2__)
-    for (int i = 0; i < num_pixels_aligned; i += 4) {
-        __m128i bgra = _mm_loadu_si128((__m128i*)(pixels + i));
-        __m128i b_mask = _mm_set1_epi32(0x000000FF);
-        __m128i r_mask = _mm_set1_epi32(0x00FF0000);
-        __m128i b = _mm_and_si128(bgra, b_mask);
-        __m128i r = _mm_and_si128(bgra, r_mask);
-        __m128i br_swapped = _mm_or_si128(_mm_slli_epi32(b, 16), _mm_srli_epi32(r, 16));
-        __m128i ga_alpha_mask = _mm_set1_epi32(0xFF00FF00);
-        __m128i ga_alpha = _mm_and_si128(bgra, ga_alpha_mask);
-        __m128i rgba = _mm_or_si128(ga_alpha, br_swapped);
-        _mm_storeu_si128((__m128i*)(pixels + i), rgba);
-    }
-#else
-    for (int i = num_pixels_aligned; i < num_pixels; ++i) {
-        uint32_t val = pixels[i];
-        pixels[i] = ((val & 0xff) << 16) | (val & 0x00ff00) | ((val & 0xff0000) >> 16) | (val & 0xff000000);
-    }
-#endif
-}
 
 int main(int argc, char **argv) {
 
@@ -74,7 +27,6 @@ int main(int argc, char **argv) {
         return -1;
     }
     printf("Successfully opened %s\n", filename);
-
 
     int32_t level = atoi(argv[2]);
     int32_t x_coord = atoi(argv[3]);
@@ -102,10 +54,9 @@ int main(int argc, char **argv) {
     uint32_t *pixels = NULL;
     assert(libisyntax_read_region(isyntax, isyntax_cache, level, x_coord, y_coord, region_width, region_height, &pixels) ==
            LIBISYNTAX_OK);
+
     // convert data to the correct pixel format (bgra->rgba).
-
     bgra_to_rgba(pixels, region_width, region_height);
-
 
     printf("Writing %s...\n", output_png);
     stbi_write_png(output_png, region_width, region_height, 4, pixels, region_width * 4);
